Extract DFP tests in test-utils
So those tests can be shared for both AdMob v20 and v19.

JIRA: DPP-3348
